Take the maven coordinates out of the dependency and into their own class
