+++ core/src/main/java/com/nikodoko/javaimports/environment/maven/MavenDependency.java
@@ -1,11 +1,8 @@
 package com.nikodoko.javaimports.environment.maven;
 
-import static com.nikodoko.javaimports.common.Utils.checkNotNull;
-
 import com.nikodoko.javaimports.common.Utils;
 import java.util.Objects;
 import java.util.Optional;
-import java.util.regex.Pattern;
 
 /** Encapsulates a Maven dependency. */
 class MavenDependency {
@@ -35,91 +32,30 @@ class MavenDependency {
       }
 
       var that = (Versionless) o;
-      return Objects.equals(this.wrapped.groupId, that.wrapped.groupId)
+      return Objects.equals(this.wrapped.groupId(), that.wrapped.groupId())
           && Objects.equals(this.wrapped.type, that.wrapped.type)
-          && Objects.equals(this.wrapped.scope, that.wrapped.scope)
-          && Objects.equals(this.wrapped.optional, that.wrapped.optional)
-          && Objects.equals(this.wrapped.artifactId, that.wrapped.artifactId);
+          && Objects.equals(this.wrapped.artifactId(), that.wrapped.artifactId());
     }
 
     @Override
     public int hashCode() {
-      return Objects.hash(
-          this.wrapped.groupId,
-          this.wrapped.artifactId,
-          this.wrapped.type,
-          this.wrapped.scope,
-          this.wrapped.optional);
+      return Objects.hash(this.wrapped.groupId(), this.wrapped.artifactId(), this.wrapped.type);
     }
 
     @Override
     public String toString() {
       return Utils.toStringHelper(this)
-          .add("groupId", this.wrapped.groupId)
-          .add("artifactId", this.wrapped.artifactId)
+          .add("groupId", this.wrapped.groupId())
+          .add("artifactId", this.wrapped.artifactId())
           .add("type", this.wrapped.type)
-          .add("scope", this.wrapped.scope)
-          .add("optional", this.wrapped.optional)
           .toString();
     }
   }
 
-  private static class Version {
-    private static final Pattern PATTERN = Pattern.compile("\\$\\{(?<parameter>\\S+)\\}");
-
-    final String value;
-    // Empty if value is not a property reference
-    final Optional<String> property;
-
-    Version(String value) {
-      this.value = value;
-      this.property = maybeExtractProperty(value);
-    }
-
-    private Optional<String> maybeExtractProperty(String value) {
-      if (value == null) {
-        return Optional.empty();
-      }
-
-      var m = PATTERN.matcher(value);
-      if (!m.matches()) {
-        return Optional.empty();
-      }
-
-      return Optional.of(m.group("parameter"));
-    }
-
-    @Override
-    public String toString() {
-      return value;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-      if (o == null) {
-        return false;
-      }
-
-      if (!(o instanceof Version)) {
-        return false;
-      }
-
-      var that = (Version) o;
-      return Objects.equals(this.value, that.value) && Objects.equals(this.property, that.property);
-    }
-
-    @Override
-    public int hashCode() {
-      return Objects.hash(this.value, this.property);
-    }
-  }
-
-  private final String groupId;
-  private final String artifactId;
-  private final Version version;
   private final String type;
-  private final String scope;
   private final boolean optional;
+  private final Optional<String> scope;
+  private final MavenCoordinates coordinates;
 
   MavenDependency(
       String groupId,
@@ -128,35 +64,17 @@ class MavenDependency {
       String type,
       String scope,
       boolean optional) {
-    checkNotNull(groupId, "maven dependency does not accept a null groupId");
-    checkNotNull(artifactId, "maven dependency does not accept a null artifactId");
-    checkNotNull(scope, "maven dependency does not accept a null scope");
-    checkNotNull(type, "maven dependency does not accept a null type");
-    this.groupId = groupId;
-    this.artifactId = artifactId;
-    this.version = new Version(version);
+    this.coordinates = new MavenCoordinates(groupId, artifactId, version, type);
     this.type = type;
-    this.scope = scope;
+    this.scope = Optional.ofNullable(scope);
     this.optional = optional;
   }
 
-  String version() {
-    return version.value;
-  }
-
-  String artifactId() {
-    return artifactId;
-  }
-
-  String groupId() {
-    return groupId;
-  }
-
   String type() {
-    return type;
+    return coordinates.type();
   }
 
-  String scope() {
+  Optional<String> scope() {
     return scope;
   }
 
@@ -164,39 +82,43 @@ class MavenDependency {
     return optional;
   }
 
-  String propertyReferencedByVersion() {
-    if (!hasPropertyReferenceVersion()) {
-      throw new IllegalStateException("Version does not reference a property: " + this);
-    }
-
-    return version.property.get();
-  }
-
-  boolean hasPropertyReferenceVersion() {
-    return version.property.isPresent();
-  }
-
-  boolean hasVersion() {
-    return version.value != null;
-  }
-
-  /**
-   * A version is assumed to be well defined if it exists and is not a property reference.
-   *
-   * <p>We don't handle cases where the version is simply invalid.
-   */
-  boolean hasWellDefinedVersion() {
-    return hasVersion() && !hasPropertyReferenceVersion();
-  }
-
   Versionless hideVersion() {
     return new Versionless(this);
   }
 
+  // All these are convenience methods calling the corresponding one in MavenCoordinates
+  // TODO: maybe get rid of them?
+  String artifactId() {
+    return coordinates.artifactId();
+  }
+
+  String groupId() {
+    return coordinates.groupId();
+  }
+
+  String version() {
+    return coordinates.version();
+  }
+
+  String propertyReferencedByVersion() {
+    return coordinates.propertyReferencedByVersion();
+  }
+
+  boolean hasPropertyReferenceVersion() {
+    return coordinates.hasPropertyReferenceVersion();
+  }
+
+  boolean hasVersion() {
+    return coordinates.hasVersion();
+  }
+
+  boolean hasWellDefinedVersion() {
+    return coordinates.hasWellDefinedVersion();
+  }
+
   @Override
   public int hashCode() {
-    return Objects.hash(
-        this.groupId, this.artifactId, this.version, this.type, this.scope, this.optional);
+    return Objects.hash(this.coordinates, this.type, this.scope, this.optional);
   }
 
   @Override
@@ -210,20 +132,16 @@ class MavenDependency {
     }
 
     MavenDependency d = (MavenDependency) o;
-    return Objects.equals(d.groupId, groupId)
-        && Objects.equals(d.artifactId, artifactId)
-        && Objects.equals(d.type, type)
+    return Objects.equals(d.type, type)
+        && Objects.equals(d.coordinates, coordinates)
         && Objects.equals(d.scope, scope)
-        && Objects.equals(d.optional, optional)
-        && Objects.equals(d.version, version);
+        && Objects.equals(d.optional, optional);
   }
 
   @Override
   public String toString() {
     return Utils.toStringHelper(this)
-        .add("groupId", groupId)
-        .add("artifactId", artifactId)
-        .add("version", version)
+        .add("coordinates", coordinates)
         .add("type", type)
         .add("scope", scope)
         .add("optional", optional)
